requirements.txt
alabaster==0.7.12
appdirs==1.4.3
APScheduler==3.5.3
attrs==18.2.0
Babel==2.6.0
backcall==0.1.0
bson==0.5.6
cached-property==1.5.1
certifi==2018.8.24
chardet==3.0.4
Click==7.0
decorator==4.3.0
defusedxml==0.5.0
docutils==0.14
future==0.17.1
geocoder==1.38.1
idna==2.7
imagesize==1.1.0
inflection==0.3.1
ipykernel==5.0.0
ipython==7.0.1
ipython-genutils==0.2.0
isodate==0.6.0
jedi==0.13.1
Jinja2==2.10
jupyter-client==5.2.3
jupyter-core==4.4.0
lockfile==0.12.2
lxml==4.2.5
MarkupSafe==1.0
numpy==1.15.2
packaging==18.0
pandas==0.23.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.5
psutil==5.4.7
ptyprocess==0.6.0
Pygments==2.2.0
pymongo==3.7.1
pyparsing==2.3.0
Pyro4==4.73
python-daemon==2.2.0
python-dateutil==2.7.3
pytz==2018.5
pyzmq==17.1.2
ratelim==0.1.6
requests==2.19.1
requests-toolbelt==0.8.0
serpent==1.27
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinx-rtd-theme==0.4.2
sphinxcontrib-websupport==1.1.0
termcolor==1.1.0
tornado==5.1.1
traitlets==4.3.2
tzlocal==1.5.1
urllib3==1.23
wcwidth==0.1.7
yappi==0.98
zeep==3.1.0
